requirements.txt
Flask==3.0.0
pandas==2.2.0
numpy==1.26.4
python-dotenv